pyproject.toml
[project]
name = "ai-trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[places]>=2.0.7",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
